requirements.txt
streamlit
numpy
opencv-python==4.1.2.30
Pillow
